Move ExerciseDetailViewModel
