+++ ArtemisKit/Sources/CourseView/ExerciseTab/ExerciseDetailView.swift
@@ -13,56 +13,6 @@ import SharedServices
 import SwiftUI
 import UserStore
 
-@Observable
-final class ExerciseDetailViewModel {
-    let courseId: Int
-    let exerciseId: Int
-
-    var exercise: DataState<Exercise>
-
-    init(courseId: Int, exerciseId: Int, exercise: DataState<Exercise>) {
-        self.courseId = courseId
-        self.exerciseId = exerciseId
-
-        self.exercise = exercise
-    }
-}
-
-extension ExerciseDetailViewModel {
-    var score: String {
-        let score = exercise.value?.baseExercise.studentParticipations?
-            .first?
-            .baseParticipation
-            .results?
-            .filter { $0.rated ?? false }
-            .max(by: { ($0.id ?? Int.min) > ($1.id ?? Int.min) })?
-            .score ?? 0
-
-        let maxPoints = exercise.value?.baseExercise.maxPoints ?? 0
-
-        return (score * maxPoints / 100).rounded().clean
-    }
-
-    var showFeedbackButton: Bool {
-        switch exercise.value {
-        case .fileUpload, .programming, .text:
-            return true
-        default:
-            return false
-        }
-    }
-
-    var isExerciseParticipationAvailable: Bool {
-        switch exercise.value {
-        case .modeling:
-            return true
-        default:
-            return false
-        }
-    }
-
-}
-
 public struct ExerciseDetailView: View {
     @EnvironmentObject var navigationController: NavigationController
